p4tc: add table create, update, delete, get, flush and dump
This commit allows users to create, update, delete, get, flush and dump
table class.

It's important to note that write operations, such as create, update
and delete, can only be made if the pipeline is not sealed.

As an example, if one were to create a table named table1 in a
pipeline named ptables, one would use the following command:

tc p4template create table/ptables/table1 keysz 32 \
keys default action metact cmd set key parse.5.1 \
postactions \
action cmd metact beq results.hit true action pipe:action ptables/drop

It also says this table will have only one key which is set by a metact
that assigns it to a specific header value extracted by the parser
(parse.5.1). Note that we are creating a key without assigning a key ID.
That is because if we don't specify one, the kernel will give us one.
In the above example, the key ID will be 1, because it's the first key
we are creating for this table class. Also note that we use the keyword
"default" when defining this key, this is to indicate that this will be
the default key for this table class. That is, if the user doesn't
specify a key when binding a table instance of this type to a filter,
the code will assume the key is the default one. In this particular
case, we don't actually need to use the default keyword when defining
the key, because we only have one. However, there are cases where this
keyword is useful. If we have more than one key and the user doesn't
use the keyword default when defining any of them, we assume the default
key is the first one. The postaction specifies what to execute after a
 match attempt has been performed in an instance of this table class. In
this particular example, if the match succeeded we'll execute action pipe
(true action pipe). Otherwise, we'll execute action ptables/drop
(action ptables/drop).

Some parameters such as tcount(number of table instances allowed for the
table class), nummasks (number of masks this table may have) and tentries
(maximum number of entries this table may have) are omitted. In
this case the default values are used, which are, respectively, 1, 8 and
256. There is also an optional argument (preactions), which will be
executed before the table key actions.

It's also important to note that the user can also specify table ID
when creating a new table class. For example:

tc p4template create table/ptables/table1 ttid 1 keysz 16 ...

The table ID is specified by using the "ttid" keyword. However,
please note that if you don't specify a table ID when creating, the
kernel will assign one for you.

If one were to update a table named table1 in a pipeline named
ptables, one would use the following command:

tc p4template update table/ptables/table1 keys action gact index 4

Note that, in the above command, we are not specifying an ID for the
key. In that scenario, the command will create a new key for this table
type, instead of updating, and will assign a new ID for the key.

If one were to delete a table named table1 from a pipeline named
ptables, one would use the following command:

tc p4template del table/ptables/table1

If one were to flush all the table classs from a pipeline named ptables,
one would use the following command:

tc p4template del table/ptables/

If one were to get a table named table1 from a pipeline named ptables,
one would use the following command:

tc p4template get table/ptables/table1

If one were to dump all the table classs from a pipeline named ptables, one
would use the following command:

tc p4template get table/ptables/

Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>