p4tc: add template actions create, update, delete, get, flush and dump
This commit allows users to create, update, delete, get, flush and dump
action templates.

Action templates work as new "dynamic" action types "on the fly". For
example, imagine that we have the following action in a P4 program:

action ipv4_forward(bit<48> dstAddr,bit<8> port)
{
     standard_metadata.egress_spec = port;
     hdr.ethernet.srcAddr = hdr.ethernet.dstAddr;
     hdr.ethernet.dstAddr = dstAddr;
     hdr.ipv4.ttl = hdr.ipv4.ttl - 1;
}

which is invoked on a table match as such:

table ipv4_lpm {
        key = {
            hdr.ipv4.dstAddr: lpm;
        }

        actions = {
            ipv4_forward;
            drop;
            NoAction;
        }

        size = 1024;
}

We don't have an equivalent built in action in TC. So our goal is
to create this action "on the fly". We do so by introducing action
templates. The description of action templates will broken down in three
parts. The first one is  templating, which deals with the action template
creation and description. The second one is runtime, which deals with
action template instantiation and binding (specifically to filters).
The last one is "Other control commands", which deals with other
important control path commands, like delete, getting, flushing and
dumping.

___TEMPLATING___

The initialisation of action templates occurs in three stages.
The first one we call "template creation".

___TEMPLATE CREATION___

In this stage we issue the template creation command which specifies the
action template name, its parameters and the parameter types.
So for the ipv4_forward action, the template creation op would look
something like this:

tc p4template create action/ptables/ipv4_forward \
  param dstAddr type mac param port type u8

Note that in the template creation op we usually just specify the action
name, the parameters and their respective types. Also see that we specify
a pipeline name during the template creation command. As an example, the
above command creates an action template that is bounded to
pipeline/program named ptables. Also, below is an example of how one
would specify an ID to the action template created in the above command.
When the create doesn't specify the action template ID, the kernel
assigns a new one for us. Also, if the action template ID specified in
the command is already in use, the kernel will reject the command.

tc p4template create action/ptables/ipv4_forward actid 1 \
  param dstAddr type mac param port type u8

___OPS_DESCRIPTION___

In the next stage (ops description), we need to specify which operations
this action uses. As example, if we were to specify the operations for
the ipv4_forward action, we'd issue the following command:

tc p4template update action/ptables/ipv4_forward
   metact \
    cmd set standard_metadata/egress_spec param/port
    cmd set hdr/ethernet/srcAddr hdr/ethernet/dstAddr
    cmd set hdr/ethernet/dstAddr param/dstAddr
    cmd decr hdr/ipv4/ttl

As you can see, we refer to the argument values in the metact use the
"action" prefix. So, for example, when referring to the argument port in a
metact, we use "action.port".

___ACTION_ACTIVATION___

Once we provided all the necessary information for the new action template,
we can go to the final stage, which is action activation. In this stage,
we activate the action template and make it available for instantiation.
To activate the action template, we issue the following command:

tc p4template update action ptables/ipv4_forward state active

After the above the command, the action is ready to be instantiated.

___RUNTIME___

This next section deals with the runtime part of action templates, which
handle action template instantiation and binding.

To instantiate a new action from a template, we use the following command:

tc actions add action ptables/ipv4_forward \
param dstAddr AA:BB:CC:DD:EE:FF param port 12 index 1

As previously stated, we refer to the action by it's "full name"
(action_name/pipeline_name). Here we are creating an instance of the
ipv4_forward action specifying as parameter values AA:BB:CC:DD:EE:FF for
dstAddr and 12 for port. We can create as many instances for action
templates as we wish.

To bind the instantiated action to a filter, you can do use the same
command used to bind ordinary actions to filter, for example:

tc filter add dev eth0 parent ffff: handle 1 prio 2 protocol ip matchall \
action ptables/ipv4_forward index 1

The above command will bind our newly instantiated action to a TC filter
atached to interface eth0 with parent qdisc ffff:, handle 1 and priority 2.
It will match all IPv4 ingres packets that arrive at this interface and
call our newly instantiated action when that happens. To learn more about
TC filters, please take a look at the TC filters man page.

___OTHER_CONTROL_COMMANDS___

This next section deals with the other control path operations we can
execute for action templates, such as delete, get, dump and flush.

As with all pipeline components, write operations to action templates, such
as create, update and delete, can only be executed if the pipeline is not
sealed.

If, after we are done with our action template we want to delete it, we
should issue the following command:

tc p4template del action/ptables/ipv4_forward

Note that we could also not specify the action name and use the ID instead,
which would transform the above command into the following:

tc p4template del action/ptables actid 1

If we had created more action templates and wanted to flush all of the
action templates from pipeline ptables, one would use the following
command:

tc p4template del action/ptables/

After creating or updating an action template, if one wishes to verify that
the action template was created correctly, one would use the following
command:

tc p4template get action/ptables/ipv4_forward

As with the del operation, when can also specify the action id instead of
the action name:

tc p4template get action/ptables actid 1

The above command will display the relevant data for the action template,
such as parameter names, types and also the associated metacts.

If one wanted to check which action templates were associated to a specific
pipeline, one could use the following command:

tc p4template get action/ptables/

Note that this command will only display the name of these action
templates. To verify their specific details, one should use the get
command, which was previously described.

Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>