tc: Increase action kind string length
Increase action kind string length from IFNAMSIZ to 64

The new dynamic actions, created by action templates, will have the
following action kind string format: "pipeline_name/act_name". IFNAMSIZ
is 16 and thus is most times to small carry a string of that format.
So, to conform to this new format, we increase the maximum name length
to account for this extra string (pipeline name) and the '/' character.

Signed-off-by: Victor Nogueira <victor@mojatatu.com>