p4tc: Add P4 types
Introduce abstraction that represents P4 program types.

We allow the user to specify bitstrings using the notation "bitX",
where X is the number of bits in a defined unsigned type and "intX", where
X is the number of bits in a defined signed type. We also allow strings
in this first commit.

XXX: This commit defines only a few P4 types that are needed at the
moment.

There will be future enhancements to add more.

Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>