p4tc: Add dynamic action's plus, sub, band, bor, bxor and concat commands
================================PLUS================================

The plus command is used to add two operands
The basic syntax for the plus command is:

cmd plus operA operB operC

The command will add operands operB and operC and store the result in
operC. That is: operA = operB + operC

operA can only be a metadatum, a key a header field.
operB and operC can only be a constant, a metadatum, a key, a header field
or a param

The following example will add metadatum mymd from pipeline ptables and
constant 16 and store the result in metadatum mymd2 of pipeline ptables:

tc p4template create action/ptables/myfunc \
   cmd plus metadata.ptables.mymd2 metadata.ptables.mymd constant.bit32.16

================================SUB================================

The sub command is used to subtract two operands
The basic syntax for the sub command is:

cmd sub operA operB operC

The command will subtract operands operB and operC and store the result in
operC. That is: operA = operB - operC

operA can only be a metadatum, a key a header field.
operB and operC can only be a constant, a metadatum, a key, a header field
or a param

The following example will subtract metadatum mymd from pipeline ptables
and constant 16 and store the result in metadatum mymd2 of pipeline ptables:

tc p4template create action/ptables/myfunc \
   cmd sub metadata.ptables.mymd2 metadata.ptables.mymd constant.bit32.16

================================CONCAT================================

The concat command is used to concat two operands and save the result to
a third one
The basic syntax for the sub command is:

cmd concat operA operB operC

The command will concat operands operB and operC and store the result in
operC. That is: operA = operB ... operC

It goes without saying that operA's size must be greater or equal to
operB's size + operC's size

operA can only be a metadatum, a key a header field.
operB and operC can only be a constant, a metadatum, a key, a header field
or a param

cmd set key.simple_l3_pna.MainControlImpl/l3_match_rx.1
hdrfield.simple_l3_pna.1.ipv4.dstAddr

The following example will concat metadatum mymd from pipeline ptables
with header field tcp.dport and store the result in metadatum mymd2 of
pipeline ptables:

tc p4template create action/ptables/myfunc cmd concat \
  metadata.ptables.mymd2 metadata.ptables.mymd hdrfield.ptables.1.tcp.dport

================================BAND================================

The band command is used to perform a binary AND operation between two operands
The basic syntax for the band command is:

cmd band operA operB operC

The command will perform the "operB AND operC" and store the result in
operC. That is: operA = operB & operC

operA can only be a metadatum, a key a header field.
operB and operC can only be a constant, a metadatum, a key, a header field
or a param

The following example will perform an AND operation of constant 16 and
mymd metadata and store the result in metadatum mymd2 of pipeline ptables:

tc p4template create action/ptables/myfunc \
   cmd band metadata.ptables.mymd2 metadata.ptables.mymd constant.bit32.16

================================BOR================================

The bor command is used to perform a binary OR operation between two operands
The basic syntax for the bor command is:

cmd bor operA operB operC

The command will perform the "operB OR operC" and store the result in
operC. That is: operA = operB | operC

operA can only be a metadatum, a key a header field.
operB and operC can only be a constant, a metadatum, a key, a header field
or a param

The following example will perform an OR operation of constant 16 and
mymd metadata and store the result in metadatum mymd2 of pipeline ptables:

tc p4template create action/ptables/myfunc \
   cmd bor metadata.ptables.mymd2 metadata.ptables.mymd constant.bit32.16

================================BXOR================================

The bxor command is used to perform a binary XOR operation between two operands
The basic syntax for the bxor command is:

cmd bxor operA operB operC

The command will perform the "operB XOR operC" and store the result in
operC. That is: operA = operB ^ operC

operA can only be a metadatum, a key a header field.
operB and operC can only be a constant, a metadatum, a key, a header field
or a param

The following example will perform an OR operation of constant 16 and
mymd metadata and store the result in metadatum mymd2 of pipeline ptables:

tc p4template create action/ptables/myfunc \
   cmd bor metadata.ptables.mymd2 metadata.ptables.mymd constant.bit32.16

Fixes: 98eb79b3f532

Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Evangelos Haleplidis <ehalep@mojatatu.com>
Signed-off-by: Evangelos Haleplidis <ehalep@mojatatu.com>
Co-developed-by: Jamal Hadi Salim <jhs@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>