p4tc: Add default hit and miss action to table class
The default hit action will be executed whenever a table lookup results in
a match and the table entry which matched has no action associated with it.

The default miss action will be executed whenever a table lookup doesn't
match any of the entries.

Both default hit and default miss are optional.

Fixes: aa1d86c234ad

Signed-off-by: Victor Nogueira <victor@mojatatu.com>