p4tc: Add 6 more optional operands to concat command
Allow user to concat 8 operands in total, instead of only two
This is intended to improve usability. Before, if the user wished to
concat, for example, 3 operands, the user would have to concat 2, store
the result in an intermediary operand and then concat the last operand
with the result of the previous concat. Now, the user just needs to call
concat passing the 3 operands (4 counting the result operand).
Like the following:

cmd concat opA opB opC opD

With this we also restructured struct p4tc_cmd_v by removing opA,
opB, opC and all variables associated with them.
We now have struct p4tc_u_interal_operand which stores all operand
information we need for parsing and assembling the netlink messages

There's also a fix to command line parsing for dynamic action commands
and another fix for verifying if the metadata operand path is complete,
that is, if it is composed of METADATA.PNAME.MNAME.

Fixes: 98eb79b3f532

Signed-off-by: Victor Nogueira <victor@mojatatu.com>