p4tc: add register operand to P4TC dynamic actions' commands
================================REGISTERS================================

Register operands are equivalent to register values in P4.
In the commands, we allow users to access registers by using their
indices, like what happends in P4 programs.

For example, if one created a P4 register issuing the following command:

tc p4template create register/ptables/register1 regid 1 type bit32
numelems 128

This means that we have a new register, associated with pipeline
ptables, with 128 elements of type bit32. This would be equivalent to
the following declaration in P4:

Register<bit<32>>(128) register1;

Like with P4, you access the registers by index. We simulate that in the
commands by using array indexing syntax. For example, if we were to
create a dynamic action where we set register1's index 0 to constant
value 12, we'd issue the following command:

tc p4template create action/ptables/test \
cmd set register.ptables.register1[0] constant.bit32.12

Note that we could index until 127 in this case because register one has
128 elements.

Register operands can be used in almost all operands that are not
specific to tables or actions (tableapply or act command). There one
excpetion which is concat. Register operands can be rvalues in the
concat command, but can't be the lvalue.

Fixes: 98eb79b3f532

Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Jamal Hadi Salim <jhs@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>