p4tc: Fix P4 types constant parsing to check value bitsz
Fix constant parsing in iproute2 to check if value fits in type bitsz.
For example, if the user tries to pass constant.bit24.0xFFFFFFFF, we
should emit an error in user space saying that this value doesn't fit in
a bit24. This is also being checked by the kernel, so those kind of
constants would make the kernel emit an error, however we should also
check this in user space to avoid the trip to the kernel.

Fixes: e9158a68329e

Signed-off-by: Victor Nogueira <victor@mojatatu.com>