p4tc: Add CRUDX permissions to table entries and const table entries
Add CRUDX permissions to table entries for data and control path.

These permissions will define if we are able to create, read, update and
delete table entries. The execute in this case specifies if we can
execute the actions associated with a table entry.

This commit also allows users to create constant table entries, by
virtue of the p4template create command. The syntax for doing so is the
following:

$TC p4template update table/ptables/cb/tname \
entry srcAddr 10.10.10.10/24 dstAddr 1.1.1.0/24 prio 17

In this command we are "updating" table cb/tname with a new entry. This
entry has as its keys srcAddr and dstAddr (both IPv4 addresses) and prio
17. As of now, we can only create one constant table entry per update
command, but we plan on batching table entries.

This update command will send a netlink response with the following
message:

pipeline id 22
    table id 1
    table name cb/tname
    key_sz 64
    max entries 256
    masks 8
    default key 1
    table entries 1
    permissions CRUD--R--X

    entry:
        table id 1
        entry priority 17
        key blob    101010a0a0a0a
        mask blob   ffffff00ffffff
        create whodunnit tc
        permissions -RUD--R--X

This message will appear if the user has tc monitor turned on

Fixes: 8c5a3389b6a0

Signed-off-by: Victor Nogueira <victor@mojatatu.com>