p4tc: Adapt dynamic action commands' to search for names and IDs
Adapt dynamic actions commands' to search for names instead of IDs to
find tables, table keys, pipelines, metadata, parsers, header fields and
registers. This will make the API more versatile and allow us to remove
the commands dependency on the introspection file, since we required
introspection to translate the names the user typed in the command to
IDs.

Fixes: 98eb79b3f532

Signed-off-by: Victor Nogueira <victor@mojatatu.com>