p4tc: Add runtime parameters to dynamic action commands
Allow dynamic action act commands to have runtime parameters.

For example, if one were to create the following action simple action,
which receives a parameter and prints its value:

$TC p4template create action/ptables/test param param1 type bit32 \
cmd print param.param1

To use this function in act command, we'd have create an instance of it
and then call that instance, for example:

$TC actions add action ptables/test param param1 type bit32 12 index 1

$TC p4template create action/ptables/test2 \
cmd act ptables.test.1

The above command will create an instance of ptables/test with index 1
and with parameter value 12

To simplify the command, this commit allows users to directly specify
the parameter value in the act command, removing the need for the
instantiation. For example, we could the same thing now using the
following command:

$TC p4template create action/ptables/test2 \
cmd act ptables.test constant.bit32.12

Fixes: 98eb79b3f532

Signed-off-by: Victor Nogueira <victor@mojatatu.com>