p4tc: Add allowed table actions list to table
Add actions list in table of actions that can be used in table entries
or as default actions. This is our attempt at emulating the table action
list in P4 programs.

We also allow flags for each of the actions in this list that specify if
the action can be added only as a table entry (tableonly), or only as a
default action (defaultonly). If no flags are specified, it is assumed
that the action can be used in both contexts.

Fixes: 16e2e009ef16

Signed-off-by: Victor Nogueira <victor@mojatatu.com>