p4tc: Add return command
Add return commands which will stop the execution of a P4TC dynamic
action and return a valid gact code (ok, drop, accept, ...).

================================RETURN================================

The return operation will stop the execution of a P4TC dynamic action and
return a valid gact code (ok, drop, accept, ...).

The return command has the following syntax:

return A

If we wanted to return a gact ok code prior to this change, we'd have to
do the following:

tc actions add action ok index 1

tc p4template create action/myprog/test actid 1 \
  cmd act kernel.gact.1

That is, to simply return an ok, we'd have to create an instance of
gact, then reference it in an act command.
Furthermore, in the data path, the code will call that instance of gact,
make it run the act callback which will, only then, return the ok code.

To simplify this, we created the return command.

So, the previous ordeal would be transformed into:

 tc p4template create action/myprog/test actid 1 \
  cmd return ok

No need to create an instance or call it in the data path, we simply
return the ok code.

Fixes: 8b83c9583cad

Signed-off-by: Victor Nogueira <victor@mojatatu.com>