p4tc: add table entry create, update, get, delete, flush and dump
This commit allows users to create, update, delete, get, flush and dump
table entries.

For example, a user issuing the following command:

    tc p4 create ptables/table/tname/tinst1 srcAddr 10.10.10.0/24 \
    dstAddr 192.168.0.0/16 prio 16

indicates a pipeline named "ptables" a table class "tname" and table
instance "tinst1".

User space tc will create a key which has a value of 0x0a0a0a00c0a00000
(10.10.10.0 concatenated with 192.168.0.0) and a mask value of
0xffffff00ffff0000 (/24 concatenated with /16) that will be sent to the
kernel. In addition a priority field of 16 is passed to the kernel.
The priority field is needed to disambiguate in case two entries
match. In that case, the kernel will choose the one with lowest priority
number.

Note that table entries can only be created once the pipeline template is
sealed.

If the user wanted to, for example, add an action to our just created
entry, they'd issue the following command:

tc p4 update ptables/table/tname/tinst1 srcAddr 10.10.10.0/24 \
dstAddr 192.168.0.0/16 prio 16 action ok

In this case, the user needs to specify the pipeline name, the table
type, the table instance, the keys and the priority, so that we can
locate the table entry.

If the user wanted to, for example, get the table entry that we just
updated, they'd issue the following command:

tc p4 get ptables/table/tname/tinst1 srcAddr 10.10.10.0/24 \
dstAddr 192.168.0.0/16 prio 16

Note that, again, we need to specify the pipeline name, the table
type, the table instance, the keys and the priority, so that we can
locate the table entry.

If the user wanted to delete the table entry we created, they'd issue the
following command:

tc p4 del ptables/table/tname/tinst1 srcAddr 10.10.10.0/24 \
dstAddr 192.168.0.0/16 prio 16

Note that, again, we need to specify the pipeline name, the table
type, the table instance, the keys and the priority, so that we can
locate the table entry.

We can also flush all the table entries from a specific table instance.
To flush the table entries of table instance named tinst1, from table
type tname and pipeline ptables, the user would issue the following
command:

tc p4 del ptables/table/tname/tinst1

We can also dump all the table entries from a specific table instance.
To dump the table entries of table instance named tinst1, from table
type tname and pipeline ptables, the user would issue the following
command:

tc p4 get ptables/table/tname/tinst1

Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>