Add prog keyword to P4 filter
Add prog keyword to P4 filter so that we can specify an eBPF program and
bind it to the filter. There are two ways of binding the eBPF parser program
to the P4 filter.

The first one is with "prog tc", which will bind the eBPF program to TC
and make it execute in the filter before any of the P4TC actions and
commands.

tc filter add dev $P0 ingress protocol any prio 1 p4 pname redirect_srcip \
         prog tc obj $PROGNAME.o section parser/tc-ingress

The second one is with "prog xdp", which will reference the eBPF program
loaded into XDP. Note that in this case the eBPF program will have to be
previously loaded into XDP before referencing it in the P4 filter.

So first we load it:

ip link set $P0 xdp obj $PROGNAME.o section xdp_ingress/xdp-ingress verbose

Then we pin it:

bpftool prog pin id $ID pin /tmp/

After that we create the P4 filter and refernce the XDP program:

$TC filter add dev $P0 ingress protocol ip prio 1 p4 pname redirect_srcip \
     prog xdp pinned /tmp/xdp_parser prog_cookie 22

Note that we also specify a "prog_cookie", which is used to verify
whether the eBPF program has executed or not before we reach the P4
classifier. The eBPF program sets this cookie by using the kfunc
bpf_p4tc_set_cookie.

Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>